Make df_appraisal function parameter

+++ css_archiving_format.py
@@ -55,54 +55,47 @@ def check_arguments(arg_list):
     return input_dir, md_path, mode, errors
 
 
-def delete_appraisal_letters(input_dir):
+def delete_appraisal_letters(input_dir, df_appraisal):
     """Deletes letters received from constituents and individual letters sent back by the office
     because they are one of the types of letters not retained for appraisal reasons"""
-
-    # Reads the appraisal delete log into a dataframe, which is in the parent folder of input_dir if it is present.
-    # If it is not, there are no files to delete.
-    try:
-        df = pd.read_csv(os.path.join(os.path.dirname(input_dir), 'appraisal_delete_log.csv'))
-    except FileNotFoundError:
-        print(f"No appraisal delete log in {os.path.dirname(input_dir)}")
-        return
 
     # Creates a file deletion log, with a header row.
     log_path = os.path.join(os.path.dirname(input_dir), f"file_deletion_log_{date.today().strftime('%Y-%m-%d')}.csv")
     file_deletion_log(log_path, None, True)
 
-    # Deletes letters received based on in_document_name.
-    # If there is a document name, it is formatted ..\documents\BlobExport\objects\filename.txt
-    in_doc_df = df.dropna(subset=['in_document_name']).copy()
-    in_doc_list = in_doc_df['in_document_name'].tolist()
-    for name in in_doc_list:
-        file_path = name.replace('..', input_dir)
-        file_path = file_path.replace('\\BlobExport', '')
-        try:
-            file_deletion_log(log_path, file_path)
-            os.remove(file_path)
-        except FileNotFoundError:
-            file_deletion_log(log_path, file_path, note='Cannot delete: FileNotFoundError')
+    # For every row in df_appraisal, delete any letter in the in_document_name and out_document_name columns.
+    # The letter path has to be reformatted to match the actual export.
+    for row in df_appraisal.itertuples():
 
-    # Deletes individual letters (not form letters) sent based on out_document_name.
-    out_doc_df = df.dropna(subset=['out_document_name']).copy()
-    out_doc_list = out_doc_df['out_document_name'].tolist()
-    for name in out_doc_list:
-        # Paths for formletters include the folder "formletter" or "form".
-        if 'form' not in name:
-            # Make an absolute path from name, which starts ..\documents or \\name-office\dos\public.
-            if name.startswith('..'):
-                file_path = name.replace('..', input_dir)
-                file_path = file_path.replace('\\BlobExport', '')
-            else:
-                file_path = re.sub('\\\\[a-z]+-[a-z]+\\\\dos\\\\public', 'documents', name)
-                file_path = input_dir + file_path
-            # Only delete if it is a file. Sometimes, out_document_name has the path to a folder instead.
-            if os.path.isfile(file_path):
+        # Deletes letters received from constituents, if the "in" column isn't blank.
+        if row.in_document_name != '':
+            name = row.in_document_name
+            file_path = name.replace('..', input_dir)
+            file_path = file_path.replace('\\BlobExport', '')
+            try:
                 file_deletion_log(log_path, file_path)
                 os.remove(file_path)
-            elif not os.path.exists(file_path):
+            except FileNotFoundError:
                 file_deletion_log(log_path, file_path, note='Cannot delete: FileNotFoundError')
+
+        # Deletes individual letters, not form letters, sent to constituents, if the "out" column isn't blank.
+        if row.out_document_name != '':
+            name = row.out_document_name
+            # Paths for formletters include the folder "formletter" or "form".
+            if 'form' not in name:
+                # Make an absolute path from name, which starts ..\documents or \\name-office\dos\public.
+                if name.startswith('..'):
+                    file_path = name.replace('..', input_dir)
+                    file_path = file_path.replace('\\BlobExport', '')
+                else:
+                    file_path = re.sub('\\\\[a-z]+-[a-z]+\\\\dos\\\\public', 'documents', name)
+                    file_path = input_dir + file_path
+                # Only delete if it is a file. Sometimes, out_document_name has the path to a folder instead.
+                if os.path.isfile(file_path):
+                    file_deletion_log(log_path, file_path)
+                    os.remove(file_path)
+                elif not os.path.exists(file_path):
+                    file_deletion_log(log_path, file_path, note='Cannot delete: FileNotFoundError')
 
 
 def file_deletion_log(log_path, file_path, header=False, note=None):
@@ -307,7 +300,7 @@ if __name__ == '__main__':
     # For preservation, deletes files for appraisal decisions.
     # It uses the log from find_appraisal_rows() to know what to delete.
     if script_mode == 'preservation':
-        delete_appraisal_letters(input_directory)
+        delete_appraisal_letters(input_directory, appraisal_df)
 
     # For access, removes rows for appraisal and columns with PII from the metadata
     # and makes a copy of the data split by congress year.
